% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_utils.R
\name{as_simple_formula2}
\alias{as_simple_formula2}
\title{Convert character variable list to formula}
\usage{
as_simple_formula2(outcome, covars)
}
\arguments{
\item{outcome}{Character string, the outcome variable.}

\item{covars}{Character vector of covariates.}
}
\value{
A formula object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated and will be removed in a future version.
}
\keyword{internal}
