% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_utils.R
\name{extract_covariates2}
\alias{extract_covariates2}
\title{Extract variable names from model terms}
\usage{
extract_covariates2(x)
}
\arguments{
\item{x}{A character vector of model terms.}
}
\value{
A character vector of unique variable names.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Takes a character vector including potentially model terms like \code{*} and \code{:} and
extracts out the individual variables.
This function is deprecated and will be removed in a future version.
}
\keyword{internal}
