% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_imputed_data}
\alias{get_imputed_data}
\title{Get Imputed Data Sets as a data frame}
\usage{
get_imputed_data(impute_obj)
}
\arguments{
\item{impute_obj}{The imputation object from which the imputed datasets are extracted.}
}
\value{
A data frame with the original subject IDs mapped and renamed.
}
\description{
This function takes an imputed dataset and a mapping variable to return a dataset
with the original IDs mapped back and renamed appropriately.
}
\examples{
# Example usage:
# result <- get_imputed_data(impute_obj)

}
