% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbmiUtils-package.R
\docType{package}
\name{rbmiUtils-package}
\alias{rbmiUtils}
\alias{rbmiUtils-package}
\title{rbmiUtils: Utility Functions to Support and Extend the 'rbmi' Package}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides utility functions that extend the capabilities of the reference-based multiple imputation package 'rbmi'. It supports clinical trial analysis workflows with functions for managing imputed datasets, applying analysis methods across imputations, and tidying results for reporting.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/openpharma/rbmiUtils}
  \item Report bugs at \url{https://github.com/openpharma/rbmiUtils/issues}
}

}
\author{
\strong{Maintainer}: Mark Baillie \email{bailliem@gmail.com} (\href{https://orcid.org/0000-0002-5618-0667}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Tobias Muetze \email{tobias.muetze@novartis.com} (\href{https://orcid.org/0000-0002-4111-1941}{ORCID})
  \item Jack Talboys \email{jack.talboys@novartis.com}
}

Other contributors:
\itemize{
  \item Lukas A. Widmer (\href{https://orcid.org/0000-0003-1471-3493}{ORCID}) [contributor]
}

}
\keyword{internal}
