% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_utils.R
\name{as_simple_formula2}
\alias{as_simple_formula2}
\title{Convert character variable list to formula}
\usage{
as_simple_formula2(outcome, covars)
}
\arguments{
\item{outcome}{Character string, the outcome variable.}

\item{covars}{Character vector of covariates.}
}
\value{
A formula object.
}
\description{
Convert character variable list to formula
}
\keyword{internal}
