% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_utils.R
\name{as_simple_formula2}
\alias{as_simple_formula2}
\title{Creates a simple formula object from a string}
\usage{
as_simple_formula2(outcome, covars)
}
\arguments{
\item{outcome}{character (length 1 vector). Name of the outcome variable}

\item{covars}{character (vector). Name of covariates}
}
\value{
A formula
}
\description{
Converts a string list of variables into a formula object
}
