% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_utils.R
\name{extract_covariates2}
\alias{extract_covariates2}
\title{Extract variable names from model terms}
\usage{
extract_covariates2(x)
}
\arguments{
\item{x}{A character vector of model terms.}
}
\value{
A character vector of unique variable names.
}
\description{
Takes a character vector including potentially model terms like \code{*} and \code{:} and
extracts out the individual variables.
}
\keyword{internal}
