% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_utils.R
\name{gcomp_responder_multi}
\alias{gcomp_responder_multi}
\title{Utility function for G-computation of a binary oucome at Multiple Visits}
\usage{
gcomp_responder_multi(data, vars, reference_levels = NULL, ...)
}
\arguments{
\item{data}{A data.frame containing all visits.}

\item{vars}{A list specifying analysis variables.}

\item{reference_levels}{Optional reference level for the treatment variable.}

\item{...}{Additional arguments passed to \code{gcomp_responder()}.}
}
\value{
A named list of estimates for each visit and treatment group.
}
\description{
Applies \code{gcomp_responder()} separately for each unique visit in the data.
}
\examples{
library(rbmi)

data("ADMI")

ADMI <- ADMI |>
  mutate(
    TRT = factor(TRT, levels = c("Placebo", "Drug A")),
    STRATA = factor(STRATA),
    REGION = factor(REGION)
  )

method <- method_bayes(
  n_samples = 100,
  control = control_bayes(warmup = 200, thin = 2)
)

ana_obj_prop <- analyse_mi_data(
  data = ADMI,
  vars = vars_binary,
  method = method,
  fun = gcomp_responder_multi,
  reference_levels = "Placebo",
  contrast = "diff",
  var_method = "Ge",
  type = "HC0"
)

print(ana_obj_prop)
pool(ana_obj_prop)

}
