% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbmiUtils-package.R
\docType{package}
\name{rbmiUtils-package}
\alias{rbmiUtils}
\alias{rbmiUtils-package}
\title{rbmiUtils: Utility functions to work with and extend the rbmi package}
\description{
This package provides additional functionality for rbmi to enable trial analysis and reporting aligned with CDISC ADaM.
}
\author{
\strong{Maintainer}: Mark Baillie \email{mark.baillie@novartis.com} (\href{https://orcid.org/0000-0002-5618-0667}{ORCID})

Authors:
\itemize{
  \item Tobias Mütze \email{tobias.muetze@novartis.com} (\href{https://orcid.org/0000-0002-4111-1941}{ORCID})
}

Other contributors:
\itemize{
  \item Jack Talboys \email{jack.talboys@novartis.com} [contributor]
  \item Lukas A. Widmer (\href{https://orcid.org/0000-0003-1471-3493}{ORCID}) [contributor]
}

}
\keyword{internal}
